+++ src/manifest.js
@@ -2,7 +2,7 @@ module.exports = {
     common: {
         name: "Quick Mining Addon",
         icon: "quick-mining.png",
-        min_engine_version: "1.20.50"
+        min_engine_version: "1.20.60"
     },
     packs: [
         {
@@ -25,7 +25,7 @@ module.exports = {
                 },
                 {
                     module_name: "@minecraft/server",
-                    version: "1.8.0-beta"
+                    version: "1.9.0-beta"
                 },
                 {
                     module_name: "@minecraft/server-ui",